Added (incomplete) source code
At present, multinomial_exp supports basic arithmetic operations and a few tentative/exploratory functions intended to express polynomials in a format with indeterminate coefficients. The related structure multinomial_evaluator is intended to facilitate computations for large/high-dimensional polynomials, but currently needs to be defined/built manually from constituent 'root' multivariate polynomials.  This project is still very much a work in progress, and very little of the code has been tested (and none of it extensively.)